requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-filter==23.5
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
kombu==5.3.5
prompt-toolkit==3.0.43
PyJWT==1.7.1
pymemcache==4.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-memcached==1.62
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.7.0
